Fixing links, config error and new requirements
Based on a new change in RTDs we need to specify the version of mkdocs to have predictable builds. If we didn't specify this we would end up on the latest version (see cfg changes with this commit as to why this isn't wanted). So we are now fixed on mkdocs 1.5.1